requirements.txt
flask
pandas
flask_cors
supabase
cryptography
google-generativeai
openai

python-dotenv
